Move muted state to Recoil Fixes #1115
