+++ client/src/hooks/useAudio.tsx
@@ -1,11 +1,12 @@
 import { useRef } from "react";
-import { useAppContext } from "../hooks/useAppContext.mjs";
+import { useRecoilValue } from "recoil";
+import { mutedState } from "../context/atoms";
 
 export type useAudioType = { play: () => void };
 
 export function useAudio(url: string): useAudioType {
   const audio = useRef(new Audio(url));
-  const { muted } = useAppContext();
+  const muted = useRecoilValue(mutedState);
 
   const play = () => {
     if (!muted) {
